pyproject.toml
[tool.poetry]
name = "de-project"
version = "0.1.0"
description = "DE course project"
authors = ["Vyacheslav Kostrov <slavkostrov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.4"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ruff = "^0.8.3"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
line-length = 120
preview = true

[tool.ruff.lint]
fixable = [
    "ALL",
]
select = [
    "A",     # flake8-builtins
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "BLE",   # flake8-blind-except
    "C4",    # flake8-comprehensions
    "C90",   # mccabe
    "COM",   # flake8-commas
    "D",     # pydocstyle
    "DTZ",   # flake8-datetimez
    "E",     # pycodestyle errors
    "ERA",   # eradicate
    "EXE",   # flake8-executable
    "F",     # pyflakes
    "FLY",   # flynt
    "FURB",  # refurb
    "I",     # isort
    "ICN",   # flake8-import-conventions
    "ISC",   # flake8-implicit-str-concat
    "N",     # pep8-naming
    "NPY",   # numpy
    "PD",    # pandas-vet
    "PERF",  # perflint
    "PIE",   # flake8-pie
    "PLC",   # pylint - convention
    "PLE",   # pylint - error
    "PLW",   # pylint - warning
    "PT",    # flake8-pytest-style
    "PTH",   # flake8-use-pathlib
    "Q",     # flake8-quotes
    "RET",   # flake8-return
    "RUF",   # Ruff-specific rules
    "S",     # flake8-bandit
    "SIM",   # flake8-simplify
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "TCH",   # flake8-type-checking
    "TID",   # flake8-tidy-imports
    "TRY",   # tryceratops
    "UP",    # pyupgrade
    "W",     # pycodestyle - warning
    "YTT",   # flake8-2020
]
ignore = [
    "D100",   # pydocstyle - undocumented-public-module
    "D104",   # pydocstyle - undocumented-public-package
    "TRY003",
    "RUF002",
    "D415",
    "DTZ011",
    "E711",
    "B026",
    "ANN",
]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["T201"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false # see: https://github.com/astral-sh/ruff/issues/8796
mark-parentheses = false    # see: https://github.com/astral-sh/ruff/issues/8796

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"
